Struct tree::set::TreeSet

pub struct TreeSet<T, C = Natural<T>> where C: Compare<T> {
    // some fields omitted
}

An ordered set based on a binary search tree.

The behavior of this set is undefined if an item's ordering relative to any other item changes while the item is in the set. This is normally only possible through Cell, RefCell, or unsafe code.

Methods

impl<T> TreeSet<T> where T: Ord

fn new() -> TreeSet<T>

Creates an empty set ordered according to the natural order of its items.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

let mut it = set.iter();
assert_eq!(it.next(), Some(&1));
assert_eq!(it.next(), Some(&2));
assert_eq!(it.next(), Some(&3));
assert_eq!(it.next(), None);

impl<T, C> TreeSet<T, C> where C: Compare<T>

fn with_cmp(cmp: C) -> TreeSet<T, C>

Creates an empty set ordered according to the given comparator.

Examples

use compare::{Compare, natural};
use tree::TreeSet;

let mut set = TreeSet::with_cmp(natural().rev());

set.insert(2);
set.insert(1);
set.insert(3);

let mut it = set.iter();
assert_eq!(it.next(), Some(&3));
assert_eq!(it.next(), Some(&2));
assert_eq!(it.next(), Some(&1));
assert_eq!(it.next(), None);

fn is_empty(&self) -> bool

Checks if the set is empty.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert!(set.is_empty());

set.insert(2);
assert!(!set.is_empty());

fn len(&self) -> usize

Returns the number of items in the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert_eq!(set.len(), 0);

set.insert(2);
assert_eq!(set.len(), 1);

fn cmp(&self) -> &C

Returns a reference to the set's comparator.

Examples

use compare::{Compare, natural};
use tree::TreeSet;

let set = TreeSet::new();
assert!(set.cmp().compares_lt(&1, &2));

let set: TreeSet<_, _> = TreeSet::with_cmp(natural().rev());
assert!(set.cmp().compares_gt(&1, &2));

fn clear(&mut self)

Removes all items from the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.len(), 3);
assert_eq!(set.iter().next(), Some(&1));

set.clear();

assert_eq!(set.len(), 0);
assert_eq!(set.iter().next(), None);

fn insert(&mut self, item: T) -> bool

Inserts an item into the set, returning true if the set did not already contain the item.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert!(!set.contains(&1));
assert!(set.insert(1));
assert!(set.contains(&1));
assert!(!set.insert(1));

fn remove<Q: ?Sized>(&mut self, item: &Q) -> bool where C: Compare<Q, T>

Removes the given item from the set, returning true if the set contained the item.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.len(), 3);
assert!(set.contains(&1));
assert!(set.remove(&1));

assert_eq!(set.len(), 2);
assert!(!set.contains(&1));
assert!(!set.remove(&1));

fn contains<Q: ?Sized>(&self, item: &Q) -> bool where C: Compare<Q, T>

Checks if the set contains the given item.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert!(!set.contains(&1));
set.insert(1);
assert!(set.contains(&1));

fn max(&self) -> Option<&T>

Returns a reference to the set's maximum item, or None if the set is empty.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert_eq!(set.max(), None);

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.max(), Some(&3));

fn min(&self) -> Option<&T>

Returns a reference to the set's minimum item, or None if the set is empty.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();
assert_eq!(set.min(), None);

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.min(), Some(&1));

fn pred<Q: ?Sized>(&self, item: &Q) -> Option<&T> where C: Compare<Q, T>

Returns a reference to the greatest item that is strictly less than the given item, or None if no such item is present in the set.

The given item need not itself be present in the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.pred(&0), None);
assert_eq!(set.pred(&1), None);
assert_eq!(set.pred(&2), Some(&1));
assert_eq!(set.pred(&3), Some(&2));
assert_eq!(set.pred(&4), Some(&3));

fn pred_or_eq<Q: ?Sized>(&self, item: &Q) -> Option<&T> where C: Compare<Q, T>

Returns a reference to the greatest item that is less than or equal to the given item, or None if no such item is present in the ste.

The given item need not itself be present in the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.pred_or_eq(&0), None);
assert_eq!(set.pred_or_eq(&1), Some(&1));
assert_eq!(set.pred_or_eq(&2), Some(&2));
assert_eq!(set.pred_or_eq(&3), Some(&3));
assert_eq!(set.pred_or_eq(&4), Some(&3));

fn succ<Q: ?Sized>(&self, item: &Q) -> Option<&T> where C: Compare<Q, T>

Returns a reference to the smallest item that is strictly greater than the given item, or None if no such item is present in the set.

The given item need not itself be present in the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.succ(&0), Some(&1));
assert_eq!(set.succ(&1), Some(&2));
assert_eq!(set.succ(&2), Some(&3));
assert_eq!(set.succ(&3), None);
assert_eq!(set.succ(&4), None);

fn succ_or_eq<Q: ?Sized>(&self, item: &Q) -> Option<&T> where C: Compare<Q, T>

Returns a reference to the smallest item that is greater than or equal to the given item, or None if no such item is present in the set.

The given item need not itself be present in the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.succ_or_eq(&0), Some(&1));
assert_eq!(set.succ_or_eq(&1), Some(&1));
assert_eq!(set.succ_or_eq(&2), Some(&2));
assert_eq!(set.succ_or_eq(&3), Some(&3));
assert_eq!(set.succ_or_eq(&4), None);

fn into_iter(self) -> IntoIter<T>

Returns an iterator that consumes the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

let mut it = set.into_iter();
assert_eq!(it.next(), Some(1));
assert_eq!(it.next(), Some(2));
assert_eq!(it.next(), Some(3));
assert_eq!(it.next(), None);

fn iter(&self) -> Iter<T>

Returns an iterator over the set.

Examples

use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

let mut it = set.iter();
assert_eq!(it.next(), Some(&1));
assert_eq!(it.next(), Some(&2));
assert_eq!(it.next(), Some(&3));
assert_eq!(it.next(), None);

fn into_range<Min: ?Sized, Max: ?Sized>(self, min: Bound<&Min>, max: Bound<&Max>) -> IntoRange<T> where C: Compare<Min, T> + Compare<Max, T>

Returns an iterator that consumes the set, yielding only those items that lie in the given range.

Examples

use std::collections::Bound::{Excluded, Unbounded};
use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.into_range(Excluded(&1), Unbounded).collect::<Vec<_>>(), [2, 3]);

fn range<Min: ?Sized, Max: ?Sized>(&self, min: Bound<&Min>, max: Bound<&Max>) -> Range<T> where C: Compare<Min, T> + Compare<Max, T>

Returns an iterator over the set's items that lie in the given range.

Examples

use std::collections::Bound::{Included, Excluded, Unbounded};
use tree::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

assert_eq!(set.range(Unbounded, Unbounded).collect::<Vec<_>>(), [&1, &2, &3]);
assert_eq!(set.range(Excluded(&1), Included(&5)).collect::<Vec<_>>(), [&2, &3]);
assert_eq!(set.range(Included(&1), Excluded(&2)).collect::<Vec<_>>(), [&1]);