Struct tree::TreeMap

pub struct TreeMap<K, V, C = Natural<K>> where C: Compare<K> {
    // some fields omitted
}

An ordered map based on a binary search tree.

Methods

impl<K, V> TreeMap<K, V> where K: Ord

fn new() -> TreeMap<K, V>

Creates an empty map ordered according to the natural order of its keys.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

let mut it = map.iter();
assert_eq!(it.next(), Some((&1, &"a")));
assert_eq!(it.next(), Some((&2, &"b")));
assert_eq!(it.next(), Some((&3, &"c")));
assert_eq!(it.next(), None);

impl<K, V, C> TreeMap<K, V, C> where C: Compare<K>

fn with_cmp(cmp: C) -> TreeMap<K, V, C>

Creates an empty map ordered according to the given comparator.

Examples

use compare::{Compare, natural};
use tree::TreeMap;

let mut map = TreeMap::with_cmp(natural().rev());

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

let mut it = map.iter();
assert_eq!(it.next(), Some((&3, &"c")));
assert_eq!(it.next(), Some((&2, &"b")));
assert_eq!(it.next(), Some((&1, &"a")));
assert_eq!(it.next(), None);

fn is_empty(&self) -> bool

Checks if the map is empty.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert!(map.is_empty());

map.insert(2, "b");
assert!(!map.is_empty());

fn len(&self) -> usize

Returns the number of entries in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.len(), 0);

map.insert(2, "b");
assert_eq!(map.len(), 1);

fn cmp(&self) -> &C

Returns a reference to the map's comparator.

Examples

use compare::{Compare, natural};
use tree::TreeMap;

let map: TreeMap<i32, &str> = TreeMap::new();
assert!(map.cmp().compares_lt(&1, &2));

let map: TreeMap<i32, &str, _> = TreeMap::with_cmp(natural().rev());
assert!(map.cmp().compares_gt(&1, &2));

fn clear(&mut self)

Removes all entries from the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.len(), 3);
assert_eq!(map.iter().next(), Some((&1, &"a")));

map.clear();

assert_eq!(map.len(), 0);
assert_eq!(map.iter().next(), None);

fn insert(&mut self, key: K, value: V) -> Option<V>

Inserts an entry into the map, returning the previous value, if any, associated with the key.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.insert(1, "a"), None);
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.insert(1, "b"), Some("a"));
assert_eq!(map.get(&1), Some(&"b"));

fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<(K, V)> where C: Compare<Q, K>

Removes and returns the entry whose key is equal to the given key, returning None if the map does not contain the key.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.len(), 3);
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.remove(&1), Some((1, "a")));

assert_eq!(map.len(), 2);
assert_eq!(map.get(&1), None);
assert_eq!(map.remove(&1), None);

fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where C: Compare<Q, K>

Checks if the map contains the given key.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert!(!map.contains_key(&1));
map.insert(1, "a");
assert!(map.contains_key(&1));

fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where C: Compare<Q, K>

Returns a reference to the value associated with the given key, or None if the map does not contain the key.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.get(&1), None);
map.insert(1, "a");
assert_eq!(map.get(&1), Some(&"a"));

fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where C: Compare<Q, K>

Returns a mutable reference to the value associated with the given key, or None if the map does not contain the key.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.get(&1), None);
map.insert(1, "a");

{
    let value = map.get_mut(&1).unwrap();
    assert_eq!(*value, "a");
    *value = "b";
}

assert_eq!(map.get(&1), Some(&"b"));

fn max(&self) -> Option<(&K, &V)>

Returns a reference to the map's maximum key and a reference to its associated value, or None if the map is empty.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.max(), None);

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.max(), Some((&3, &"c")));

fn max_mut(&mut self) -> Option<(&K, &mut V)>

Returns a reference to the map's maximum key and a mutable reference to its associated value, or None if the map is empty.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.max(), None);

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let max = map.max_mut().unwrap();
    assert_eq!(max, (&3, &mut "c"));
    *max.1 = "cc";
}

assert_eq!(map.max(), Some((&3, &"cc")));

fn min(&self) -> Option<(&K, &V)>

Returns a reference to the map's minimum key and a reference to its associated value, or None if the map is empty.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.min(), None);

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.min(), Some((&1, &"a")));

fn min_mut(&mut self) -> Option<(&K, &mut V)>

Returns a reference to the map's minimum key and a mutable reference to its associated value, or None if the map is empty.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();
assert_eq!(map.min(), None);

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let min = map.min_mut().unwrap();
    assert_eq!(min, (&1, &mut "a"));
    *min.1 = "aa";
}

assert_eq!(map.min(), Some((&1, &"aa")));

fn pred<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where C: Compare<Q, K>

Returns a reference to the greatest key that is strictly less than the given key and a reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.pred(&0), None);
assert_eq!(map.pred(&1), None);
assert_eq!(map.pred(&2), Some((&1, &"a")));
assert_eq!(map.pred(&3), Some((&2, &"b")));
assert_eq!(map.pred(&4), Some((&3, &"c")));

fn pred_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(&K, &mut V)> where C: Compare<Q, K>

Returns a reference to the greatest key that is strictly less than the given key and a mutable reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let pred = map.pred_mut(&2).unwrap();
    assert_eq!(pred, (&1, &mut "a"));
    *pred.1 = "aa";
}

assert_eq!(map.pred(&2), Some((&1, &"aa")));

fn pred_or_eq<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where C: Compare<Q, K>

Returns a reference to the greatest key that is less than or equal to the given key and a reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.pred_or_eq(&0), None);
assert_eq!(map.pred_or_eq(&1), Some((&1, &"a")));
assert_eq!(map.pred_or_eq(&2), Some((&2, &"b")));
assert_eq!(map.pred_or_eq(&3), Some((&3, &"c")));
assert_eq!(map.pred_or_eq(&4), Some((&3, &"c")));

fn pred_or_eq_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(&K, &mut V)> where C: Compare<Q, K>

Returns a reference to the greatest key that is less than or equal to the given key and a mutable reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let pred_or_eq = map.pred_or_eq_mut(&1).unwrap();
    assert_eq!(pred_or_eq, (&1, &mut "a"));
    *pred_or_eq.1 = "aa";
}

{
    let pred_or_eq = map.pred_or_eq_mut(&4).unwrap();
    assert_eq!(pred_or_eq, (&3, &mut "c"));
    *pred_or_eq.1 = "cc";
}

assert_eq!(map.pred_or_eq(&1), Some((&1, &"aa")));
assert_eq!(map.pred_or_eq(&4), Some((&3, &"cc")));

fn succ<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where C: Compare<Q, K>

Returns a reference to the smallest key that is strictly greater than the given key and a reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.succ(&0), Some((&1, &"a")));
assert_eq!(map.succ(&1), Some((&2, &"b")));
assert_eq!(map.succ(&2), Some((&3, &"c")));
assert_eq!(map.succ(&3), None);
assert_eq!(map.succ(&4), None);

fn succ_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(&K, &mut V)> where C: Compare<Q, K>

Returns a reference to the smallest key that is strictly greater than the given key and a mutable reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let succ = map.succ_mut(&2).unwrap();
    assert_eq!(succ, (&3, &mut "c"));
    *succ.1 = "cc";
}

assert_eq!(map.succ(&2), Some((&3, &"cc")));

fn succ_or_eq<Q: ?Sized>(&self, key: &Q) -> Option<(&K, &V)> where C: Compare<Q, K>

Returns a reference to the smallest key that is greater than or equal to the given key and a reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

assert_eq!(map.succ_or_eq(&0), Some((&1, &"a")));
assert_eq!(map.succ_or_eq(&1), Some((&1, &"a")));
assert_eq!(map.succ_or_eq(&2), Some((&2, &"b")));
assert_eq!(map.succ_or_eq(&3), Some((&3, &"c")));
assert_eq!(map.succ_or_eq(&4), None);

fn succ_or_eq_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<(&K, &mut V)> where C: Compare<Q, K>

Returns a reference to the smallest key that is greater than or equal to the given key and a mutable reference to its associated value, or None if no such key is present in the map.

The given key need not itself be present in the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

{
    let succ_or_eq = map.succ_or_eq_mut(&0).unwrap();
    assert_eq!(succ_or_eq, (&1, &mut "a"));
    *succ_or_eq.1 = "aa";
}

{
    let succ_or_eq = map.succ_or_eq_mut(&3).unwrap();
    assert_eq!(succ_or_eq, (&3, &mut "c"));
    *succ_or_eq.1 = "cc";
}

assert_eq!(map.succ_or_eq(&0), Some((&1, &"aa")));
assert_eq!(map.succ_or_eq(&3), Some((&3, &"cc")));

fn into_iter(self) -> IntoIter<K, V>

Returns an iterator that consumes the map.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

let mut it = map.into_iter();
assert_eq!(it.next(), Some((1, "a")));
assert_eq!(it.next(), Some((2, "b")));
assert_eq!(it.next(), Some((3, "c")));
assert_eq!(it.next(), None);

fn iter(&self) -> Iter<K, V>

Returns an iterator over the map's entries with immutable references to the values.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert(2, "b");
map.insert(1, "a");
map.insert(3, "c");

let mut it = map.iter();
assert_eq!(it.next(), Some((&1, &"a")));
assert_eq!(it.next(), Some((&2, &"b")));
assert_eq!(it.next(), Some((&3, &"c")));
assert_eq!(it.next(), None);

fn iter_mut(&mut self) -> IterMut<K, V>

Returns an iterator over the map's entries with mutable references to the values.

Examples

use tree::TreeMap;

let mut map = TreeMap::new();

map.insert("b", 2);
map.insert("a", 1);
map.insert("c", 3);

let mut i = 1;

for (_, value) in map.iter_mut() {
    assert_eq!(i, *value);
    *value *= 2;
    i += 1;
}

assert_eq!(map["a"], 2);
assert_eq!(map["b"], 4);
assert_eq!(map["c"], 6);

fn range<Min: ?Sized, Max: ?Sized>(&self, min: Bound<&Min>, max: Bound<&Max>) -> Range<K, V> where C: Compare<Min, K> + Compare<Max, K>

Returns an iterator over the map's entries whose keys lie in the given range with immutable references to the values.

Examples

use std::collections::Bound::{Included, Excluded, Unbounded};
use tree::TreeMap;

let mut map = TreeMap::new();

map.insert("b", 2);
map.insert("a", 1);
map.insert("c", 3);

assert_eq!(map.range(Unbounded, Unbounded).collect::<Vec<_>>(),
    [(&"a", &1), (&"b", &2), (&"c", &3)]);
assert_eq!(map.range(Excluded(&"a"), Included(&"f")).collect::<Vec<_>>(),
    [(&"b", &2), (&"c", &3)]);
assert_eq!(map.range(Included(&"a"), Excluded(&"b")).collect::<Vec<_>>(),
    [(&"a", &1)]);

fn range_mut<Min: ?Sized, Max: ?Sized>(&mut self, min: Bound<&Min>, max: Bound<&Max>) -> RangeMut<K, V> where C: Compare<Min, K> + Compare<Max, K>

Returns an iterator over the map's entries whose keys lie in the given range with mutable references to the values.

Examples

use std::collections::Bound;
use tree::TreeMap;

let mut map = TreeMap::new();

map.insert("b", 2);
map.insert("a", 1);
map.insert("c", 3);

let mut i = 1;

for (_, value) in map.range_mut(Bound::Unbounded, Bound::Excluded(&"c")) {
    assert_eq!(i, *value);
    *value *= 2;
    i += 1;
}

assert_eq!(map["a"], 2);
assert_eq!(map["b"], 4);
assert_eq!(map["c"], 3);